requirements.txt
pyspark
python-dateutil
pathlib
kafka-python
zipfile37